continue solidity

+++ src/solidityByExample/if_else.sol
@@ -3,7 +3,7 @@
 pragma solidity ^0.8.24;
 
 contract If_Else {
-    function foo(uint256 x) public  pure returns(uint){
+    function foo(uint256 x) public pure returns (uint256) {
         if (x < 0) {
             return 0;
         } else if (x < 20) {
@@ -13,7 +13,7 @@ contract If_Else {
         }
     }
 
-    function foo2(uint256 x) public pure returns(uint) {
+    function foo2(uint256 x) public pure returns (uint256) {
         x < 20 ? 0 : 1;
     }
 }
